requirements.txt
streamlit
pillow
python-dotenv
streamlit-option-menu
openstacksdk
mongoengine
python-dotenv
htbuilder